+++ NEWS.md
@@ -1,3 +1,5 @@
+# srvyr (development version)
+
 # srvyr 1.1.0
 * Uses the new quantile functions provided in version 4.1 of the survey package. The old survey quantile functions can be accessed with `survey_old_quantile()` and `survey_old_median()`
 * Adds a new function `interact` that makes it easier to calculate proportions among interacted groups