+++ NEWS.md
@@ -1,4 +1,6 @@
-# srvyr (development version)
+# srvyr 1.1.1
+* Add function `cur_svy_wts()` to access the survey weights (#136, #139, thanks @ray-p144 and @bschneidr)
+* Allow access to survey context functions like `cur_svy()` and `cur_svy_wts()` in `mutate` and `filter` (#138, #139, thanks @ray-p144 and @bschneidr)
 * Improve behavior of `interact()` when using `cascade()`(#133, thanks @szimmer)
 * Fix a bug with non-standard names of grouping variables (like `1234`) in cascade (#132, thanks @szimmer)